added ARE YOU SURE code to deleteServerFromPrefs

+++ Nagios Manager/AppDelegate.applescript
@@ -184,10 +184,11 @@ script AppDelegate
 			set my theServerName to x's serverName --grab the server name
 			set my theServerAPIKey to x's serverAPIKey --grab the server key
 			set my theServerURL to x's serverURL --grab the server URL
-			tell my userTable to setDoubleAction:"deleteSelectedUsers:" --this lets a doubleclick work as well as clicking the delete button. We may remove this
-			--because it could be dangrous
 			my getServerUsers:(missing value) --use missing value because we have to pass something. in ths case, the ASOC version of nil
 		end if
+		
+		tell my userTable to setDoubleAction:"deleteSelectedUsers:" --this lets a doubleclick work as well as clicking the delete button. We may remove this
+		--because it could be dangerous
           
           --set the initial state and enabled of the checkboxes
           my canSeeAllObjects's setEnabled:true
@@ -287,28 +288,39 @@ script AppDelegate
 	end loadServersFromPrefs:
 	
 	on deleteServerFromPrefs:sender --this was deleteServer:
-		my theServerTableController's remove:(theServerTableController's selectedObjects()) --deletes the selected row right out of the controller
-		--my god, this was so easy once I doped it out
-		my theSMSettingsList's removeAllObjects() --blow out theSMSettingsList
-		my theSMSettingsList's addObjectsFromArray:(theServerTableController's arrangedObjects()) --rebuild it from theServerTableController
-		--this way, at least in here, theServerTableController and theSettingsList are ALWAYS in sync and that's IMPORTANT.
+		--the ARE YOU SURE YOU WANT TO DO THIS??? warning
+		set theSelection to theServerTableController's selectedObjects() as record
+		set theServerNameToBeDeleted to theSMTableServerName of theSelection
 		
-		set theServerTableControllerObjectCount to my theServerTableController's arrangedObjects()'s |count|() --get number of objects left in
-		--the controller. Vertical bars are necessary because "count" is also an AppleScript keyword, so the bars keep it from being the AS count
-		--and instead use it as the ASOC count, which is what we want.
+		set theDeleteServerAlertButtonRecord to display alert "You are about to delete " & theServerNameToBeDeleted & " from the saved list of servers. \r\rTHIS IS NOT UNDOABLE, ARE YOU SURE?" as critical buttons {"OK","Cancel"} default button "Cancel" giving up after 90
+		set theDeleteServerButton to button returned of theDeleteServerAlertButtonRecord
 		
-		if theServerTableControllerObjectCount = 0 then --if the list is empty (we just deleted the last thing) then we'll call deleteAllServersFromPrefs and
-			--save time since that's what deleteAllServersFromPrefs does, if you think about it
-			my deleteAllServersFromPrefs:(missing value) --this handles explicitly clearing the defaults AND hasDefaults for us.
-			--technically that may not be necessary, but this way we KNOW.
+		if theDeleteServerButton is "OK" then
+			my theServerTableController's remove:(theServerTableController's selectedObjects()) --deletes the selected row right out of the controller
+			--my god, this was so easy once I doped it out
+			my theSMSettingsList's removeAllObjects() --blow out theSMSettingsList
+			my theSMSettingsList's addObjectsFromArray:(theServerTableController's arrangedObjects()) --rebuild it from theServerTableController
+			--this way, at least in here, theServerTableController and theSettingsList are ALWAYS in sync and that's IMPORTANT.
 			
-		else --so we have entries in the array, let's write that to disk
-		--what's interesting is that we already have theServerTableController and theSettingsList in the desired state, so this gets SIMPLE
-			set my theSMDefaultsExist to true --since we're writing a setting, we want to set this correctly.
+			set theServerTableControllerObjectCount to my theServerTableController's arrangedObjects()'s |count|() --get number of objects left in
+			--the controller. Vertical bars are necessary because "count" is also an AppleScript keyword, so the bars keep it from being the AS count
+			--and instead use it as the ASOC count, which is what we want.
 			
-			theDefaults's setObject:my theSMSettingsList forKey:"serverSettingsList" --write the new settings list to defaults
-			theDefaults's setBool:my theSMDefaultsExist forKey:"hasDefaults" --setting hasDefaults to true (1), this way we avoid the
-			--"but I thought it was okay" problem. We don't think we know what hasDefaults is on exit, we KNOW
+			if theServerTableControllerObjectCount = 0 then --if the list is empty (we just deleted the last thing) then we'll call deleteAllServersFromPrefs and
+				--save time since that's what deleteAllServersFromPrefs does, if you think about it
+				my deleteAllServersFromPrefs:(missing value) --this handles explicitly clearing the defaults AND hasDefaults for us.
+				--technically that may not be necessary, but this way we KNOW.
+				
+				else --so we have entries in the array, let's write that to disk
+				--what's interesting is that we already have theServerTableController and theSettingsList in the desired state, so this gets SIMPLE
+				set my theSMDefaultsExist to true --since we're writing a setting, we want to set this correctly.
+				
+				theDefaults's setObject:my theSMSettingsList forKey:"serverSettingsList" --write the new settings list to defaults
+				theDefaults's setBool:my theSMDefaultsExist forKey:"hasDefaults" --setting hasDefaults to true (1), this way we avoid the
+				--"but I thought it was okay" problem. We don't think we know what hasDefaults is on exit, we KNOW
+			end if
+		else if theDeleteServerButton is "Cancel" then
+			return
 		end if
 		
 	end deleteServerFromPrefs:
